refactor 1099r controller out into shared controller
